♻️ Moved 'call' command to own controller
