+++ pyproject.toml
@@ -104,3 +104,9 @@ markers = [
 [tool.isort]
 profile = "black"
 src_paths = ["src", "tests"]
+
+[tool.autoflake]
+in_place = true
+remove_all_unused_imports = true
+ignore_pass_after_docstring = true
+ignore_pass_statements = true
